Implemented client command raw packet creator and socket watchdog
Command class is for putting command together and gain raw packet data
according to protocol.

Client class is now inherits Tread, a watchdog is firing "info" command
every 1 s. The result will be displayed in a corner, TBD. Also, result
can be used for detecting server loss.

Probably, there's a better solution to detect server loss,
I've tried select(), but it was not satisfying. Whatever.

Upon server disconnect, it's not enough just to exit(), in the main
thread input() must be aborted (otherwise it will keep wrong
terminal setting after exiting). That's why fatal() sends SIGINT,
if requested, it is caught by input() and it performs proper cleanup.

Why we stick to readline.input(), if we have such problems with it?
Because it implements CLI history and in-line editing.

TODO for client side:
- display info polled in every 1 s,
- implement some CLI parser and
- response parser.