Partially implemented HashTable.performSet().
HashTable.findItem() - not implemented yet - will return the address
of a pointer to item. This pointer can be:
- an item of HashTable.items[],
- or a reference in another HashItem's HashItem.next.

Update is simple, needs only to update HashItem.valueData and valueLength.
It's been implemented.

Delete is trickier: the deleted item's next will be the replacement
of the deleted item, even if it's a NULL. That's why HashTable.findItem()
returns HashItem**, it makes possible to manipulate the pointer to
the deleted item. Other words, delete is removing an element from
the linked list starts at each hash table slot.