Command "set" is now implemented.
The most important method which is also needed by the "set"
command is
 HashItem** HashTable_findItem()
It returns the address of the pointer where the item
resides. It can be a hashtable slot or a "next" property
in a SlotItem. Its pair is the
 HashItem** HashItem_getNextPtr()
which returns the address of the "next" property of the
HashItem.

Known issue: half baked "get" command crashes the client.