Search (incl. providing result) implemented. SearchOptions renamed to Search.
Search contains the search options (incl. search/count mode),
and now a dynamic array of result Items. (That's why it gets renamed.)

As far as the server is single-threaded, it is not necessary to
take ownership of Items (in cases when the Item is changing or
gets deleted during a search is in progress).

Search function is implemented in HashTable.search(), which
checks match with HashItem.searchKey() and HashItem.SearchValue(),
uses Search.addResult() to store results, which Command.universalSearch()
converts into result format by using Packet.appendChunk().

There was a design dilemma with search result memory allocation.

One of the options is to use two rounds. First round counts the
results, so memory allocation can take proper number of results,
then second round collects and stores the matching objects.
This implies that
- full-search search should be performed twice on items
  (one for count, another for collect results rount),
  which has high cost (comparing twice), or
- first round should mark matching HashItem objects, so
  second round has not to perform expensive compare method
  again, only check this flag.
Marking SearchItems is okay in single-thread, but it causes
kind of mixed responsibility, HashItem depends heavy on search
(it should only depend on HashTable).

Second option is to allocate memory for potential largest number
of results, then perform search in a single round.

I've voted on the second option:
- Only pointers should be allocated for Items, which is not
  a significant amount of memory, 100k result items requires
  400K/800K of memory.
- I assume, it's better to allocate 800K memory, than scan 100k
  items twice.
- Also, I've tried to optimize a little: Search.addResult only
  allocates memory for result list only at first hit (lazy).
  This makes possible to decrease the number of allocated items
  decrease by number of items skipped until then. It has no
  effect when the 1st item is already a match, but hopefully it
  performs well when there are only 1-2 results, or, even better
  where there is no result at all (in this case, no memory will
  be allocated for result list).
With this optimization attempt, second option looks way better
than the first one.